Minor fixes
I also removed the greedy solution from the array, as it caused the block schedule system to add all of the children into the array. This caused duplications which doubled the size of the search area. This can be added later if we ensure that the path to the greedy solution is not added later.